test: move the scratch-worktree safety suite into gitscratch
The three properties that make it acceptable to dry-run against a
developer's real repository — `rebase.updateRefs` never moving a real
branch ref, detaching so branches checked out in other worktrees still
replay, and teardown never touching an unrelated worktree whose
directory is temporarily missing — are all guarantees of the harness,
not of grist's ranking. Now that the harness lives in `gitscratch`, the
tests have to live there too. Left in `grist`, the next tool built on
the harness (`grime`, `grind`) would inherit the safety and none of the
proof, and `cargo test -p gitscratch` would report green on a crate
whose entire reason to exist is untested.

Each test now drives `Scratch` directly instead of `Simulator`: create
a scratch at the base, then per branch check it out detached through
the scratch's own git runner and replay it — exactly what
`Simulator::land` does minus the squash. The `--detach` is spelled out
in the test helper on purpose, because it is itself one of the guards
under test. Test 2 lost its `ranked.len()` assertion along with
`Simulator`; it now asserts the conflicting replay really reported
conflicts, so it cannot pass by replaying nothing. Tests 1 and 3 scope
the `Scratch` to a block so its `Drop` teardown has definitely run
before the assertions — previously that happened for free inside
`Simulator::score`.

`gitscratch` dev-depends on itself to turn on its own `testing`
feature, so plain `cargo test -p gitscratch` compiles the suite without
a `--features` flag; a safety suite that only ran under
`cargo test --workspace` would silently skip on a targeted run.

All three guards were mutation-verified after the move: dropping
`rebase.updateRefs=false` fails test 1, dropping `--detach` fails test
2, and adding a repo-wide `worktree prune` to teardown fails test 3.

Refs #327
